Securisation des modif/suppr des liens et membres

+++ suppr_liens.php
@@ -1,8 +1,10 @@
 <?php
 include_once("Url.php");
+include_once("Membre.php");
 
-//TODO verifier que le lien appartient bien au pseudo connecté
-
-Url::supprimerUrl($_GET['id']);
+// verifier que le lien appartient bien au pseudo connecté ou est admin :
+if( Membre::estAdmin($_SESSION['connex_active']) || Membre::getIdFromPseudo($_SESSION['connex_active']) == Url::getAuthor($_GET['id']) )
+	Url::supprimerUrl($_GET['id']);
+	
 header("Location: mes_liens.php");
 ?>
