+++ suppr_liens2.php
@@ -1,8 +1,10 @@
 <?php
 include_once("Url.php");
+include_once("Membre.php");
 
-//TODO verifier qui le lien appartient bien au pseudo connecté
+// verifier que le lien appartient bien au pseudo connecté ou est admin :
+if( Membre::estAdmin($_SESSION['connex_active']) )
+	Url::supprimerUrl($_GET['id']);
 
-Url::supprimerUrl($_GET['id']);
 header("Location: administration.php");
 ?>
